Updated field

+++ ZenDev.Api/ApiModels/MindfulnessApiModel.cs
@@ -4,7 +4,7 @@
     {
         public long MindfulnessId { get; set; }
         public long TotalPoints { get; set; }
-        public long TotalMinutes { get; set; }
+        public Double TotalMinutes { get; set; }
         public DateTimeOffset? LastUpdate { get; set; }
         public long UserId { get; set; }
     }
